requirements.txt
flask
gunicorn
psycopg2-binary
icalendar
bcrypt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
Pillow
python-dateutil
dateparser
APScheduler
